+++ R/import_from_gis.R
@@ -5,11 +5,12 @@
 #'     Must match the database name in `query`.
 #' @param uid character, user ID. Default is `getOption("councilR.uid")`
 #' @param pwd character, user password. Default is `getOption("councilR.pwd")`.
+#' @param .quiet logical, whether to print time elapsed message.
 #'
 #' @note See `vignette("Options")` to review package options.
 #'     You must be set up with the appropriate database drivers
 #'     to use this function.
-#'     **Windows** users need ODBC with Microsoft SQL. 
+#'     **Windows** users need ODBC with Microsoft SQL.
 #'       Contact IS support for ODBC installation.
 #'     **Mac** users need `unixodbc` and `freetds`.
 #'       See instructions in the [onboarding guide](https://furry-adventure-596f3adb.pages.github.io/database-connections.html)
@@ -35,13 +36,16 @@
 import_from_gis <- function(query,
                             dbname = "GISLibrary",
                             uid = getOption("councilR.uid"),
-                            pwd = getOption("councilR.pwd")) {
+                            pwd = getOption("councilR.pwd"),
+                            .quiet = FALSE) {
   purrr::map(
     c(query, dbname, uid, pwd),
     rlang:::check_string
   )
 
-  tictoc::tic()
+  if (.quiet == FALSE) {
+    tictoc::tic()
+  }
   if (DBI::dbCanConnect(odbc::odbc(),
     # driver = "FreeTDS",
     dbname,