+++ tests/testthat/test-import_from_gis.R
@@ -1,8 +1,7 @@
 # Can import_from_gis() actually import from the GISLibrary database?
 
-# skip on GH Actions, CRAN
+# skip on GH Actions
 testthat::skip_on_ci()
-testthat::skip_on_cran()
 
 # skip if not connected to VPN
 testthat::skip_if(
@@ -16,9 +15,10 @@ testthat::skip_if(httr2::secret_has_key("QUHBRb_yoy2RRj59qno8NVXA7mW402xkins"))
 testthat::test_that("airports spatial dataset", {
   airport <- import_from_gis(
     uid = httr2::secret_decrypt("QUHBRb_yoy2RRj59qno8NVXA7mW402xkins", "COUNCILR_KEY"),
-    pwd = httr2::secret_decrypt("YXJZM6G8aHF-KTXwGiUx4Zm04qmRfERs", "COUNCILR_KEY"),
+    pwd = httr2::secret_decrypt("rQHk4S39pjfJ6yoKWUUNpQUDk2i9XA3d", "COUNCILR_KEY"),
     query = "GISLibrary.dbo.AIRPORTS",
-    dbname = "GISLibrary"
+    dbname = "GISLibrary",
+    .quiet = TRUE
   )
 
   # test that all airports are included